+++ packages/rehype-mathjax/lib/adaptor.browser.js
