+++ packages/rehype-mathjax/lib/adaptor.js
