+++ packages/rehype-mathjax/lib/core.js
