+++ packages/rehype-mathjax/lib/input.js
