+++ packages/rehype-mathjax/lib/output-chtml.js
