+++ packages/rehype-mathjax/lib/output-svg.js
