+++ packages/rehype-mathjax/lib/renderer.js
